requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-arm64
# created-by: conda 24.11.0
aom=3.9.1=h7bae524_0
appdirs=1.4.4=pyh9f0ad1d_0
appnope=0.1.4=pyhd8ed1ab_0
asttokens=2.4.1=pyhd8ed1ab_0
brotli=1.1.0=hd74edd7_2
brotli-bin=1.1.0=hd74edd7_2
brotli-python=1.1.0=py311h3f08180_2
bzip2=1.0.8=h99b78c6_7
c-ares=1.34.3=h5505292_1
ca-certificates=2024.8.30=hf0a4a13_0
cairo=1.18.0=hb4a6bf7_3
certifi=2024.8.30=pyhd8ed1ab_0
cffi=1.17.1=py311h3a79f62_0
charset-normalizer=3.4.0=pyhd8ed1ab_0
click=8.1.7=unix_pyh707e725_0
colorama=0.4.6=pyhd8ed1ab_0
comm=0.2.2=pyhd8ed1ab_0
contourpy=1.3.1=py311h210dab8_0
cpython=3.11.10=py311hd8ed1ab_3
cycler=0.12.1=pyhd8ed1ab_0
dav1d=1.2.1=hb547adb_0
debugpy=1.8.9=py311h155a34a_0
decorator=5.1.1=pyhd8ed1ab_0
docker-pycreds=0.4.0=py_0
exceptiongroup=1.2.2=pyhd8ed1ab_0
executing=2.1.0=pyhd8ed1ab_0
ffmpeg=7.1.0=gpl_h92f8dbd_105
filelock=3.16.1=pyhd8ed1ab_0
font-ttf-dejavu-sans-mono=2.37=hab24e00_0
font-ttf-inconsolata=3.000=h77eed37_0
font-ttf-source-code-pro=2.038=h77eed37_0
font-ttf-ubuntu=0.83=h77eed37_3
fontconfig=2.15.0=h1383a14_1
fonts-conda-ecosystem=1=0
fonts-conda-forge=1=0
fonttools=4.55.0=py311h4921393_0
freetype=2.12.1=hadb7bae_2
fribidi=1.0.10=h27ca646_0
fsspec=2024.10.0=pyhff2d567_0
gdk-pixbuf=2.42.12=h7ddc832_0
giflib=5.2.2=h93a5062_0
gitdb=4.0.11=pyhd8ed1ab_0
gitpython=3.1.43=pyhd8ed1ab_0
gmp=6.3.0=h7bae524_2
gmpy2=2.1.5=py311hb5ce3a2_2
graphite2=1.3.13=hebf3989_1003
h2=4.1.0=pyhd8ed1ab_0
harfbuzz=9.0.0=h997cde5_1
hdf5=1.14.4=nompi_ha698983_103
hpack=4.0.0=pyh9f0ad1d_0
hyperframe=6.0.1=pyhd8ed1ab_0
icu=75.1=hfee45f7_0
idna=3.10=pyhd8ed1ab_0
imageio=2.36.0=pyh12aca89_1
imath=3.1.12=h025cafa_0
importlib-metadata=8.5.0=pyha770c72_0
ipykernel=6.29.5=pyh57ce528_0
ipython=8.29.0=pyh707e725_0
jasper=4.2.4=h6c4e4ef_0
jedi=0.19.2=pyhff2d567_0
jinja2=3.1.4=pyhd8ed1ab_0
joblib=1.4.2=pyhd8ed1ab_0
jupyter_client=8.6.3=pyhd8ed1ab_0
jupyter_core=5.7.2=pyh31011fe_1
kiwisolver=1.4.7=py311h2c37856_0
krb5=1.21.3=h237132a_0
lame=3.100=h1a8c8d9_1003
lcms2=2.16=ha0e7c42_0
lerc=4.0.0=h9a09cb3_0
libabseil=20240722.0=cxx17_hf9b8971_1
libaec=1.1.3=hebf3989_0
libasprintf=0.22.5=h8414b35_3
libass=0.17.3=hf20b609_0
libblas=3.9.0=19_osxarm64_openblas
libbrotlicommon=1.1.0=hd74edd7_2
libbrotlidec=1.1.0=hd74edd7_2
libbrotlienc=1.1.0=hd74edd7_2
libcblas=3.9.0=19_osxarm64_openblas
libcurl=8.10.1=h13a7ad3_0
libcxx=19.1.4=ha82da77_0
libdeflate=1.22=hd74edd7_0
libedit=3.1.20191231=hc8eb9b7_2
libev=4.33=h93a5062_2
libexpat=2.6.4=h286801f_0
libffi=3.4.2=h3422bc3_5
libgettextpo=0.22.5=h8414b35_3
libgfortran=5.0.0=13_2_0_hd922786_3
libgfortran5=13.2.0=hf226fd6_3
libglib=2.82.2=h07bd6cf_0
libhwloc=2.11.2=default_hbce5d74_1001
libiconv=1.17=h0d3ecfb_2
libintl=0.22.5=h8414b35_3
libjpeg-turbo=3.0.0=hb547adb_1
liblapack=3.9.0=19_osxarm64_openblas
liblapacke=3.9.0=19_osxarm64_openblas
libnghttp2=1.64.0=h6d7220d_0
libopenblas=0.3.24=openmp_hd76b1f2_0
libopencv=4.10.0=headless_py311hc436ff1_10
libopenvino=2024.4.0=hbfeda7a_2
libopenvino-arm-cpu-plugin=2024.4.0=hbfeda7a_2
libopenvino-auto-batch-plugin=2024.4.0=hf276634_2
libopenvino-auto-plugin=2024.4.0=hf276634_2
libopenvino-hetero-plugin=2024.4.0=h03892cd_2
libopenvino-ir-frontend=2024.4.0=h03892cd_2
libopenvino-onnx-frontend=2024.4.0=h7f5a098_2
libopenvino-paddle-frontend=2024.4.0=h7f5a098_2
libopenvino-pytorch-frontend=2024.4.0=h5833ebf_2
libopenvino-tensorflow-frontend=2024.4.0=h9d544f2_2
libopenvino-tensorflow-lite-frontend=2024.4.0=h5833ebf_2
libopus=1.3.1=h27ca646_1
libpng=1.6.44=hc14010f_0
libprotobuf=5.28.2=h8f0b736_0
librsvg=2.58.4=h40956f1_0
libsodium=1.0.20=h99b78c6_0
libsqlite=3.47.0=hbaaea75_1
libssh2=1.11.1=h9cc3647_0
libtiff=4.7.0=hfce79cd_1
libtorch=2.5.1=cpu_generic_hed481e3_3
libuv=1.49.2=h7ab814d_0
libvpx=1.14.1=h7bae524_0
libwebp=1.4.0=h54798ee_0
libwebp-base=1.4.0=h93a5062_0
libxcb=1.17.0=hdb1d25a_0
libxml2=2.13.5=hbbdcc80_0
libzlib=1.3.1=h8359307_2
llvm-openmp=19.1.4=hdb05f8b_0
markupsafe=3.0.2=py311h56c23cb_0
matplotlib=3.9.2=py311ha1ab1f8_2
matplotlib-base=3.9.2=py311hbe3227e_2
matplotlib-inline=0.1.7=pyhd8ed1ab_0
mpc=1.3.1=h8f1351a_1
mpfr=4.2.1=hb693164_3
mpmath=1.3.0=pyhd8ed1ab_0
munkres=1.0.7=py_1
ncurses=6.5=h7bae524_1
nest-asyncio=1.6.0=pyhd8ed1ab_0
networkx=3.4.2=pyh267e887_2
nomkl=1.0=h5ca1d4c_0
numpy=2.1.3=py311h649a571_0
opencv=4.10.0=headless_py311hc0d1824_10
openexr=3.3.2=h04410fd_0
openh264=2.5.0=h774163f_0
openjpeg=2.5.2=h9f1df11_0
openssl=3.4.0=h39f12f2_0
packaging=24.2=pyhff2d567_1
pandas=2.2.3=pypi_0
pango=1.54.0=h9ee27a3_2
parso=0.8.4=pyhd8ed1ab_0
pcre2=10.44=h297a79d_2
pexpect=4.9.0=pyhd8ed1ab_0
pickleshare=0.7.5=py_1003
pillow=11.0.0=py311h3894ae9_0
pip=24.3.1=pyh8b19718_0
pixman=0.43.4=hebf3989_0
platformdirs=4.3.6=pyhd8ed1ab_0
pqdict=1.4.0=pypi_0
prompt-toolkit=3.0.48=pyha770c72_0
protobuf=5.28.2=py311h6885ffc_0
psutil=6.1.0=py311hae2e1ce_0
pthread-stubs=0.4=hd74edd7_1002
ptyprocess=0.7.0=pyhd3deb0d_0
pugixml=1.14=h13dd4ca_0
pure_eval=0.2.3=pyhd8ed1ab_0
py-opencv=4.10.0=headless_py311he1e16fa_10
pycparser=2.22=pyhd8ed1ab_0
pygments=2.18.0=pyhd8ed1ab_0
pyparsing=3.2.0=pyhd8ed1ab_1
pysocks=1.7.1=pyha2e5f31_6
python=3.11.10=hc51fdd5_3_cpython
python-dateutil=2.9.0.post0=pyhff2d567_0
python_abi=3.11=5_cp311
pytorch=2.5.1=cpu_generic_py311h9d776f9_3
pytz=2024.2=pypi_0
pyyaml=6.0.2=py311h460d6c5_1
pyzmq=26.2.0=py311h730b646_3
qhull=2020.2=h420ef59_5
readline=8.2=h92ec313_1
requests=2.32.3=pyhd8ed1ab_0
scikit-learn=1.5.2=py311h9e23f0f_1
scipy=1.14.1=py311hf1db568_1
sentry-sdk=2.19.0=pyhd8ed1ab_0
setproctitle=1.3.4=py311h917b07b_0
setuptools=75.6.0=pyhff2d567_1
six=1.16.0=pyh6c4a22f_0
sleef=3.7=h8391f65_2
smmap=5.0.0=pyhd8ed1ab_0
snappy=1.2.1=hd02b534_0
soundfile=0.12.1=pypi_0
stack_data=0.6.2=pyhd8ed1ab_0
svt-av1=2.3.0=hf24288c_0
sympy=1.13.3=pyh2585a3b_104
tbb=2022.0.0=h0cbf7ec_0
threadpoolctl=3.5.0=pyhc1e730c_0
tk=8.6.13=h5083fa2_1
torchaudio=2.5.1=py311_cpu
torchvision=0.20.1=py311_cpu
tornado=6.4.2=py311h917b07b_0
tqdm=4.67.1=pyhd8ed1ab_0
traitlets=5.14.3=pyhd8ed1ab_0
typing_extensions=4.12.2=pyha770c72_0
tzdata=2024.2=pypi_0
unicodedata2=15.1.0=py311hae2e1ce_1
urllib3=2.2.3=pyhd8ed1ab_0
wandb=0.18.7=py311h3ff9189_0
wcwidth=0.2.13=pyhd8ed1ab_0
wfdb=4.1.2=pypi_0
wheel=0.45.1=pyhd8ed1ab_0
x264=1!164.3095=h57fd34a_2
x265=3.5=hbc6ce65_3
xorg-libxau=1.0.11=hd74edd7_1
xorg-libxdmcp=1.1.5=hd74edd7_0
xxhash=3.5.0=pypi_0
xz=5.2.6=h57fd34a_0
yacs=0.1.8=pyhd8ed1ab_0
yaml=0.2.5=h3422bc3_2
zeromq=4.3.5=hc1bb282_7
zipp=3.21.0=pyhd8ed1ab_0
zlib=1.3.1=h8359307_2
zstandard=0.23.0=py311ha60cc69_1
zstd=1.5.6=hb46c0d2_0
